+++ src/Core/Logistics.Domain/Constraints/LoadConsts.cs
@@ -2,8 +2,8 @@
 
 public static class LoadConsts
 {
-    public const double MinDeliveryCost = 0;
-    public const double MaxDeliveryCost = 1000000d;
+    public const decimal MinDeliveryCost = 0;
+    public const decimal MaxDeliveryCost = 1000000;
     public const int AddressLength = 256;
     public const int NameLength = 64;
 }
