+++ src/Shared/Logistics.Models/IGrossChart.cs
@@ -3,7 +3,7 @@
 public interface IGrossChart
 {
     public DateTime Date { get; set; }
-    public double DriverShare { get; set; }
-    public double Gross { get; set; }
+    public decimal DriverShare { get; set; }
+    public decimal Gross { get; set; }
     public double Distance { get; set; }
 }
